only one model extractor and tweak on export_tool function
